- Implemented PodcastTask DTO for structured podcast processing. - Developed Podigee podcast processor with authentication, metadata fetching, and analytics transformation. - Created Spotify podcast processor with endpoint management and async processing. - Refactored existing task management to utilize new processors. - Enhanced error handling and logging for better traceability.
