+++ connector_manager/manager/processors/__init__.py
@@ -1,0 +1,1 @@
+# Empty init file for processors package
