+++ connector_manager/requirements.txt
@@ -20,6 +20,7 @@ tenacity==8.1.0
 toml==0.10.2
 urllib3==1.26.12
 huey==2.5.0
+aiohttp==3.9.1
 pytest>=8.4.0
 pytest-cov>=6.2.0 --prefer-binary
 requests==2.28.1
